start object oriented approach
