+++ lib/gams-view.js
@@ -36,7 +36,6 @@ export default class StructureView {
 				refShown: false,
 				control: null,
 				controlShown: false,
-				cursorListener: [],
 				noTagHint: null,
 				lastFile: null,
 				viewShow: false,
@@ -89,6 +88,7 @@ export default class StructureView {
 
 						this.cursorListener.add(editorSubscriptions)
 					}))
+					console.log(this.cursorListener)
 				},
 
 				updateValues(activeEditor, clickedSym) {
@@ -120,6 +120,7 @@ export default class StructureView {
 				},
 				destroy() {
 					this.cursorListener.dispose()
+					console.log(this.cursorListener)
 				}
 			}
 		})
