+++ lib/linter.js
@@ -13,7 +13,7 @@ shell.config.execPath = shell.which('node')
 // Compile the GAMS code, read the listing file, return the file contents
 // and delete the file when finished reading
 function compile (config, callback) {
-	// check if exp_starter.gms exists (GGIG project), else just check input file
+	// check if exp_starter.gms exists (or other GGIG project entry file), else just check input file
 	let ggigParams = ''
 	findUp(['exp_starter.gms', 'capmod.gms', 'com_.gms'], {cwd: config.directory}).then((expStarter) => {
 		if (expStarter) {
@@ -47,9 +47,7 @@ function compile (config, callback) {
 
 		// cd to given gams working directory and execute the gams compilation
 		shell.cd(config.directory)
-		console.log(`${config.gamsexe} ${config.filePath} action=c o=${nullOutput} ${tagsParam} fErr=${config.scratchdir}${path.sep}${randStr}.err -workdir="${config.directory}" -curDir="${config.directory}" ${ggigParams}`)
 		shell.exec(`${config.gamsexe} ${config.filePath}  action=c o=${nullOutput} ${tagsParam} fErr=${config.scratchdir}${path.sep}${randStr}.err -workdir="${config.directory}" -curDir="${config.directory}" ${ggigParams}`, function(code, stdout, stderr) {
-			console.log(stdout)
 			if (stderr) {
 				throw new Error(stderr)
 			}
@@ -59,8 +57,10 @@ function compile (config, callback) {
 				fs.unlink(`${config.scratchdir}${path.sep}${randStr}.err`, (err) => {
 					if (err) throw err
 				})
-				if (config.createTags) createTags(`${config.scratchdir}${path.sep}${randStr}.ref`, config.projectPath)
-				createTree(`${config.scratchdir}${path.sep}${randStr}.ref`)
+				if (config.createTags) {
+					createTags(`${config.scratchdir}${path.sep}${randStr}.ref`, config.projectPath)
+					createTree(`${config.scratchdir}${path.sep}${randStr}.ref`)
+				}
 				return callback(data.toString('utf8'))
 			})
 		})
